feat(wallet): add multi-wallet support and rename to WalletContextProvider
