+++ frontend/src/app/config/WalletContextProvider.tsx
@@ -1,35 +1,40 @@
 "use client"
 
-import { FC, useMemo } from "react"
+import { FC, useMemo, ReactNode } from "react"
 import {
   ConnectionProvider,
   WalletProvider
 } from "@solana/wallet-adapter-react"
 import { WalletModalProvider } from "@solana/wallet-adapter-react-ui"
+
+// Adapters you actually want to support
 import {
   PhantomWalletAdapter,
-  SolflareWalletAdapter
+  SolflareWalletAdapter,
+  LedgerWalletAdapter,
+  TorusWalletAdapter
 } from "@solana/wallet-adapter-wallets"
+import { GlowWalletAdapter } from "@solana/wallet-adapter-glow"
 
-import { BackpackWalletAdapter } from "@solana/wallet-adapter-backpack"
+// Mainnet RPC endpoint
+const MAINNET_ENDPOINT = "https://api.mainnet-beta.solana.com"
 
-export const WalletContextProvider: FC<{ children: React.ReactNode }> = ({
+export const WalletContextProvider: FC<{ children: ReactNode }> = ({
   children
 }) => {
-  // You can change the RPC endpoint later if needed
-  const endpoint = "https://api.devnet.solana.com"
-
   const wallets = useMemo(
     () => [
       new PhantomWalletAdapter(),
+      new GlowWalletAdapter(),
       new SolflareWalletAdapter(),
-      new BackpackWalletAdapter()
+      new LedgerWalletAdapter(),
+      new TorusWalletAdapter()
     ],
     []
   )
 
   return (
-    <ConnectionProvider endpoint={endpoint}>
+    <ConnectionProvider endpoint={MAINNET_ENDPOINT}>
       <WalletProvider wallets={wallets} autoConnect>
         <WalletModalProvider>{children}</WalletModalProvider>
       </WalletProvider>
