+++ frontend/src/app/page.tsx
@@ -1,7 +1,5 @@
 "use client"
 
-import { WalletMultiButton } from "@solana/wallet-adapter-react-ui"
-
 export default function HomePage() {
   return (
     <main className='flex flex-col items-center justify-center h-screen'>
