+++ frontend/src/components/WalletConnect.tsx
@@ -11,9 +11,7 @@ export function WalletConnect() {
   const [balance, setBalance] = useState<number | null>(null)
   const [mounted, setMounted] = useState(false)
 
-  useEffect(() => {
-    setMounted(true)
-  }, [])
+  useEffect(() => setMounted(true), [])
 
   useEffect(() => {
     if (!mounted || !connected || !publicKey) return
@@ -33,21 +31,15 @@ export function WalletConnect() {
     return () => clearInterval(interval)
   }, [mounted, connected, publicKey, connection])
 
-  if (!mounted) return null // prevent server/client mismatch
+  if (!mounted) return null
 
   return (
     <div className='flex items-center gap-4 justify-end p-4'>
       <WalletMultiButton />
-      {connected && publicKey ? (
-        <div className='text-sm font-mono'>
-          <div>
-            Address: {publicKey.toBase58().slice(0, 6)}...
-            {publicKey.toBase58().slice(-4)}
-          </div>
-          <div>SOL: {balance?.toFixed(4)}</div>
+      {connected && balance !== null && (
+        <div className='text-sm font-mono text-gray-300'>
+          SOL: {balance.toFixed(4)}
         </div>
-      ) : (
-        <div className='text-sm text-gray-400'>Wallet not connected</div>
       )}
     </div>
   )
